refactor: Player
 User와 Dealer의 extends 와 implement 구조에서
 implement만 하도록 변경
